Implement 'Boutique Inventory' exercise - notes
Exercism had the following notes about this
submission:

Essential
Use the Enum.reduce function in the total_quantity
function to practice. It's the best fitting Enum
function for this task.
(
  Not sure I agree, my solution is pretyt tight:
 `Map.values(item[:quantity_by_size]) |> Enum.sum()`
)

Essential
Use either the Map.new/2 function or the Enum.into/3
function in the increase_quantity function. They're
the best fitting functions for this task and they
don't require the usage of the Enum.map/2 function.

Essential
Use the Enum.sort_by function in the sort_by_price
function to practice. It's the best fitting Enum
function for this task.