Implement 'Community Garden' exercise (accepted)
Not sure if I structured the Agent data as well as
I could have. Needed to maintain a set of stucts
(%Plot{id: <int>, registered_to: <string>}), and the
next plot ID to use.

Originally, the data structure I used was

%{next_id: n, plots: [%Plot{}, ...]}.

The problem was that I needed to be able to find/update plots
based on the id. There didn't seem to be a clean way
to select a plot :plot_id without polling all the records.
So I modifed the data structure to this instead:

%{next_id: n, plots: %{n => %Plot{}, ...}}.

It meant replicating the plot ID val, once within the
struct, then again as a key to the struct in the map.