Implement 'Lucas Numbers' exercise (accepted)
Exercism accepted my solution, and I think it's
clear and effecient, but I suspect that it wasn't
really what they intended for this exercise.

The overview discussed Streams, but I didn't use them.
My only Enum calls were to count/1 and reverse/1, so
there really wasn't anywhere to optimize with a
Stream.