Refactor 'Lucas Numbers' exercise
Should have looked deeper at the Stream module
before submitting the last go.

Using Stream.unfold/2 |> Enum.take/1 provided a
short, simple 2 line solution.

unfold is a weird one, but perfect for this job.
The accumulator is a two element tuple, in which the
first element is the value in the sequence, and the second
is the value to be passed to the function to generate the
next value.