Refactor 'Pangram' exercise
On the first pass of the exercise, I piped the
sentence through a series of transformations to
produce an alphabetical char list of the uniq
characters in the sentence. Then checked for a
match to the alphabet ('abc...z').

This time after transforming the sentence to a
lower case charlist, I just did an Enum.all? with the
rang ?a..?z to verify that each character was in the charlist.
Slighty simpler.