Implement 'Simple Linked List' (accepted)
Perhaps I just have too much C on the brain, but this didn't really
feel very Linked-Listy to me. When I think of a linked list, I think
of a collection of pairs: data & a pointer to the next data. This was
more like, "Make an Elixir Tuple act like a Elixir List."

Maybe I missed the point of the exercise.